Remove React strict mode to prevent buggy behavior in development environment
In version 18 of React, StrictMode causes useEffect to run twice in
development mode which leads to unexpected behavior. This behavior
was added by the React team in anticipation of a future update which
will add reusable state. However, this causes undesired bugs during
the development process such as decrementing a state variable twice
instead of once. For this reason, StrictMode was removed from this
project.